修改

+++ router/institutions/institutions.js
@@ -28,33 +28,33 @@ MongoClient.connect(url, {
 				for(var i = 0; i < result.length; i++) {
 					result[i].key = i
 				}
-				logs.toTree(result).then((children)=>{
+				logs.toTree(result).then((children) => {
 					var data = {
-									code: 200,
-									data: children,
-									total: num,
-								}
-								res.jsonp(data)
+						code: 200,
+						data: children,
+						total: num,
+					}
+					res.jsonp(data)
 				})
-								
+
 			})
 		})
 	})
-	router.get('/choose',function(req,res,next){
+	router.get('/choose', function(req, res, next) {
 		dbo.collection("instruction").find({}).toArray(function(err, result) {
-				for(var i = 0; i < result.length; i++) {
-					result[i].title = result[i].name
-					result[i].value = result[i]._id
+			for(var i = 0; i < result.length; i++) {
+				result[i].title = result[i].name
+				result[i].value = result[i]._id
+			}
+			logs.toTree(result).then((children) => {
+				var data = {
+					code: 200,
+					data: children,
 				}
-				logs.toTree(result).then((children)=>{
-					var data = {
-									code: 200,
-									data: children,
-								}
-								res.jsonp(data)
-				})
-								
+				res.jsonp(data)
 			})
+
+		})
 	})
 	router.post('/add', function(req, res, next) {
 		var myobj = req.body
@@ -90,24 +90,53 @@ MongoClient.connect(url, {
 	router.post('/delete', function(req, res, next) { //删除字典，可批量
 		var arr = []
 		var ress = res
+		var flags = true
 		var idarr = req.body.id.split(',')
-		idarr.forEach(v => {
-			v = ObjectId(v)
-			arr.push(v)
+		var pList = []
+		for(var i = 0; i < idarr.length; i++) {
+			(function(i) {
+				pList.push(new Promise(function(resolve, reject) {
+					dbo.collection("instruction").find({
+						"pid": idarr[i]
+					}).toArray(function(err, result) {
+
+						if(result.length != 0) {
+							flags = false
+						}
+						resolve()
+					})
+
+				}))
+			})(i);
+			idarr[i] = ObjectId(idarr[i])
+			arr.push(idarr[i])
+		}
+		Promise.all(pList).then(function(res) {
+			console.log(flags)
+			if(flags) {
+				var myobj = {
+					"_id": {
+						$in: arr
+					}
+				};
+				dbo.collection("instruction").deleteMany(myobj, function(err, obj) {
+					if(err) throw err;
+					var data = {
+						code: 200,
+						msg: '删除成功'
+					}
+					ress.jsonp(data);
+				});
+			} else {
+				var data = {
+					code: 202,
+					msg: '删除的对象含有子级,请先删除子级',
+				}
+				ress.jsonp(data);
+			}
+
 		})
-		var myobj = {
-			"_id": {
-				$in: arr
-			}
-		};
-		dbo.collection("instruction").deleteMany(myobj, function(err, obj) {
-			if(err) throw err;
-			var data = {
-				code: 200,
-				msg: '删除成功'
-			}
-			ress.jsonp(data);
-		});
+
 	})
 })
